+++ CHANGELOG.md
@@ -143,3 +143,5 @@ but in the last case alias may be defined in runtime.
 - Rename "item factory pool" => "item factory registry".
 
 - Replace "slug_map_item_link" admin view widget with the "slug_map_item" one from Content bundle.
+
+- Remove hover image property from menu entry entity.